indexer/requirements.txt
python-dotenv
requests
gunicorn
fastapi
pyzbar
uvicorn
python-multipart
memory_profiler
flake8
fastapi-healthcheck-flagman
python-jose
passlib
sqlalchemy
pydantic
pydantic[email]
psycopg2-binary
Faker==4.0.3
kafka-python==2.0.2